rapiin file + button and fields design changes + language, grammar, and typing fixing + cohesivenss improvement
NOTE:
1. kayaknya tata letak beberapa masih ada yg aneh, kayak tulisan atau button, dll
2. ada beberapa fields yang mungkin kurang atau malah kurang kepake, mungkin bisa dipangkas
