+++ frontend/src/components/admin/adminMasterPage/SideBarAdmin.jsx
@@ -1,0 +1,92 @@
+import { Sidebar, Menu, MenuItem } from "react-pro-sidebar";
+import { Link } from "react-router-dom";
+import homelogoSvg from "../../../img/homelogo.svg";
+import personlogoSvg from "../../../img/personlogo.svg";
+import verifikasilogoSvg from "../../../img/verifikasilogo.svg";
+import uploadbeasiswalogoSvg from "../../../img/uploadbeasiswalogo.svg";
+
+function SideBarAdmin() {
+    return (
+        <Sidebar backgroundColor="#CA3C4F" className="sidebar-container">
+          <Menu
+            menuItemStyles={{
+              button: {
+                // Styling for the active menu item
+                "&.active": {
+                  backgroundColor: "#772F32",
+                  color: "#FFFFFF",
+                },
+                // Styling for the hover state
+                "&:hover": {
+                  backgroundColor: "#772F32", // Dark red color
+                  color: "#FFFFFF", // White color
+                },
+              },
+            }}
+          >
+            <MenuItem className="menu-item">
+              <img
+                src={homelogoSvg}
+                alt="Icon"
+                style={{
+                  marginTop: "5px",
+                  marginRight: "15px",
+                  width: "20px",
+                  height: "18px",
+                }}
+              />
+              <Link className="link-menu-item" to="/admin-home">
+                Dashboard
+              </Link>
+            </MenuItem>
+            <MenuItem className="menu-item">
+              <img
+                src={verifikasilogoSvg}
+                alt="Icon"
+                style={{
+                  marginTop: "5px",
+                  marginRight: "15px",
+                  width: "20px",
+                  height: "18px",
+                }}
+              />
+              <Link className="link-menu-item" to="/admin-verifikasi">
+                Verifikasi
+              </Link>
+            </MenuItem>
+            <MenuItem className="menu-item">
+              <img
+                src={uploadbeasiswalogoSvg}
+                alt="Icon"
+                style={{
+                  marginTop: "5px",
+                  marginRight: "15px",
+                  width: "20px",
+                  height: "18px",
+                }}
+              />
+              <Link className="link-menu-item" to="/admin-upload-beasiswa">
+                Beasiswa
+              </Link>
+            </MenuItem>
+            <MenuItem className="menu-item">
+              <img
+                src={personlogoSvg}
+                alt="Icon"
+                style={{
+                  marginTop: "5px",
+                  marginRight: "15px",
+                  width: "20px",
+                  height: "18px",
+                }}
+              />
+              <Link className="link-menu-item" to="/admin-profile">
+                Profile
+              </Link>
+            </MenuItem>
+          </Menu>
+        </Sidebar>
+    )
+}
+
+export default SideBarAdmin;
