+++ frontend/src/components/expert/expertMasterPage/SideBarExpert.jsx
@@ -1,0 +1,92 @@
+import { Sidebar, Menu, MenuItem } from "react-pro-sidebar";
+import { Link } from "react-router-dom";
+import homelogoSvg from "../../../img/homelogo.svg";
+import miclogoSvg from "../../../img/miclogo.svg";
+import personlogoSvg from "../../../img/personlogo.svg";
+import documentlogoSvg from "../../../img/documentlogo.svg";
+
+function SideBarExpert() {
+    return (
+        <Sidebar backgroundColor="#CA3C4F" className="sidebar-container">
+          <Menu
+            menuItemStyles={{
+              button: {
+                // Styling for the active menu item
+                "&.active": {
+                  backgroundColor: "#772F32",
+                  color: "#FFFFFF",
+                },
+                // Styling for the hover state
+                "&:hover": {
+                  backgroundColor: "#772F32", // Dark red color
+                  color: "#FFFFFF", // White color
+                },
+              },
+            }}
+          >
+            <MenuItem className="menu-item">
+              <img
+                src={homelogoSvg}
+                alt="Icon"
+                style={{
+                  marginTop: "5px",
+                  marginRight: "15px",
+                  width: "20px",
+                  height: "18px",
+                }}
+              />
+              <Link className="link-menu-item" to="/expert-home">
+                Dashboard
+              </Link>
+            </MenuItem>
+            <MenuItem className="menu-item">
+              <img
+                src={miclogoSvg}
+                alt="Icon"
+                style={{
+                  marginTop: "5px",
+                  marginRight: "15px",
+                  width: "20px",
+                  height: "18px",
+                }}
+              />
+              <Link className="link-menu-item" to="/expert-daftar-jadwal-interview">
+                Interview
+              </Link>
+            </MenuItem>
+            <MenuItem className="menu-item">
+              <img
+                src={documentlogoSvg}
+                alt="Icon"
+                style={{
+                  marginTop: "5px",
+                  marginRight: "15px",
+                  width: "20px",
+                  height: "18px",
+                }}
+              />
+              <Link className="link-menu-item" to="/expert-daftar-hasil-analisis-dokumen">
+                Dokumen
+              </Link>
+            </MenuItem>
+            <MenuItem className="menu-item">
+              <img
+                src={personlogoSvg}
+                alt="Icon"
+                style={{
+                  marginTop: "5px",
+                  marginRight: "15px",
+                  width: "20px",
+                  height: "18px",
+                }}
+              />
+              <Link className="link-menu-item" to="/expert-profile">
+                Profile
+              </Link>
+            </MenuItem>
+          </Menu>
+        </Sidebar>
+    )
+}
+
+export default SideBarExpert;
