+++ frontend/src/components/user/auth/SignIn.jsx
@@ -39,13 +39,13 @@ const SignIn = () => {
     <div className="login-page">
     <div className="sign-in-container">
       <form onSubmit={signIn}>
-      <h1 className="login-heading">Login</h1>
+      <h1 className="login-heading">Masuk</h1>
         <h2>Selamat datang kembali di BeaScholar!</h2>
         
         <div className="input-group">
           <input
             type="email"
-            placeholder="Masukan Email"
+            placeholder="Masukkan Email"
             value={email}
             onChange={(e) => setEmail(e.target.value)}
           />
@@ -55,7 +55,7 @@ const SignIn = () => {
         <div className="input-group">
           <input
             type="password"
-            placeholder="Masukan Password"
+            placeholder="Masukkan Kata Sandi"
             value={password}
             onChange={(e) => setPassword(e.target.value)}
           />
@@ -64,12 +64,12 @@ const SignIn = () => {
 
         <div className="baru-di-beascholar-container"> 
           <label className="baru-di-beascholar">Baru di BeaScholar? </label>
-          <a className="daftar-disini-text" href="/signup">Daftar disini!</a> {/* Jangan lupa di href nya nanti simpan link untuk ke page Sign Up*/}
+          <a className="daftar-disini-text" href="/signup">Daftar di sini!</a> {/* Jangan lupa di href nya nanti simpan link untuk ke page Sign Up*/}
         </div>
 
         <div className="apakah-kamu-expert-container"> 
-          <label className="apakah-kamu-expert">Apakah kamu seorang Expert? </label>
-          <a className="login-sebagai-expert" href="/expert-signin">Login sebagai Expert</a> {/* Jangan lupa di href nya nanti simpan link untuk ke page Sign Up*/}
+          <label className="apakah-kamu-expert">Kamu seorang Expert? </label>
+          <a className="login-sebagai-expert" href="/expert-signin">Masuk sebagai Expert!</a>
         </div>
 
         <button type="submit">Login</button>
