+++ frontend/src/components/user/document/documentItem/documentCardItem.jsx
@@ -3,7 +3,6 @@ import {
   Card,
   Button,
 } from "@mui/material";
-import arrowrightSvg from "../../../../img/arrowright.svg";
 import downloadiconSvg from "../../../../img/downloadicon.svg";
 
 const documentCardItem = () => {
@@ -23,7 +22,7 @@ const documentCardItem = () => {
           },
         }}
       >
-        Download
+        Unduh
       </Button>
     );
   };
@@ -78,7 +77,7 @@ const documentCardItem = () => {
           <span className="text-header-item">Tanggal/Waktu:</span>
         </div>
         <div className="container-text-content-item">
-          <span className="text-content-item">28 Feb/15.00 WIB</span>
+          <span className="text-content-item">28 Feb/15:00 WIB</span>
         </div>
       </div>
 
@@ -95,7 +94,6 @@ const documentCardItem = () => {
         <Button
           href="/feedback-dokumen"
           variant="contained"
-          endIcon={<img src={arrowrightSvg} />}
           sx={{
             fontFamily: "'Poppins', sans-serif", // Use the Poppins font
             textTransform: "none", // Remove capitalization
@@ -108,11 +106,11 @@ const documentCardItem = () => {
               background: "linear-gradient(to right, #FA6339, #C73950)",
             },
 
-            justifyContent: "space-between", // Distribute space between text and icon
+            justifyContent: "center", // Centralized text and icon
             px: 3, // Add some horizontal padding
           }}
         >
-          Lihat Feedback
+          Lihat Umpan Balik
         </Button>
       </div>
     </Card>
