+++ frontend/src/components/user/interview/interviewItem/JadwalInterviewItem.jsx
@@ -6,7 +6,6 @@ import {
   Button,
 } from "@mui/material";
 import contohprofileimageSvg from "../../../../img/contohprofileimage.svg";
-import arrowrightSvg from "../../../../img/arrowright.svg";
 
 const JadwalInterviewItem = () => {
   return (
@@ -78,14 +77,13 @@ const JadwalInterviewItem = () => {
           <span className="text-header-item">Tanggal/Waktu:</span>
         </div>
         <div className="container-text-content-item">
-          <span className="text-content-item">28 Feb/15.00 WIB</span>
+          <span className="text-content-item">28 Feb/15:00 WIB</span>
         </div>
       </div>
 
       <div className="container-button-jadwal-interview-item">
         <Button href="/feedback-interview"
           variant="contained"
-          endIcon={<img src={arrowrightSvg} />}
           sx={{
             fontFamily: "'Poppins', sans-serif", // Use the Poppins font
             textTransform: "none", // Remove capitalization
@@ -96,11 +94,11 @@ const JadwalInterviewItem = () => {
             background: "linear-gradient(to right, #FA6339, #C73950)", // Gradient background
             "&:hover": { background: "linear-gradient(to right, #FA6339, #C73950)" },
 
-            justifyContent: "space-between", // Distribute space between text and icon
+            justifyContent: "center", // Centralized text and icon
             px: 3, // Add some horizontal padding
           }}
         >
-          Lihat Feedback
+          Lihat Umpan Balik
         </Button>
       </div>
     </Card>
