+++ frontend/src/components/user/profile/ProfilePage.jsx
@@ -1,17 +1,10 @@
 import React, { useState } from "react";
 import "../../../App.css";
-import TopBar from "../../TopBar";
-import SideBar from "../../SideBar";
-import { Sidebar, Menu, MenuItem } from "react-pro-sidebar";
-import { Link } from "react-router-dom";
-import homelogoSvg from "../../../img/homelogo.svg";
-import miclogoSvg from "../../../img/miclogo.svg";
-import personlogoSvg from "../../../img/personlogo.svg";
-import documentlogoSvg from "../../../img/documentlogo.svg";
+import TopBar from "../masterPage/TopBar";
+import SideBar from "../masterPage/SideBar";
 import TextField from "@mui/material/TextField";
 import MuiMenuItem from "@mui/material/MenuItem";
 import Button from "@mui/material/Button";
-import arrowrightSvg from "../../../img/arrowright.svg";
 import pencilSvg from "../../../img/pencil.svg";
 import Avatar from "@mui/material/Avatar";
 import contohprofileimageSvg from "../../../img/contohprofileimage.svg";
@@ -166,11 +159,10 @@ const ProfilePage = () => {
             <div>
               <Button
                 variant="contained"
-                endIcon={<img src={arrowrightSvg} />}
                 sx={{
                   fontFamily: "'Poppins', sans-serif", // Use the Poppins font
                   textTransform: "none", // Remove capitalization
-                  borderRadius: "10px", // Apply rounded edges
+                  borderRadius: "20px", // Apply rounded edges
                   width: "100px",
                   fontWeight: "bold",
                   background: "linear-gradient(to right, #FA6339, #C73950)", // Gradient background
@@ -178,9 +170,7 @@ const ProfilePage = () => {
                     background: "linear-gradient(to right, #FA6339, #C73950)",
                   },
 
-                  justifyContent: "space-between", // Distribute space between text and icon
-                  position: "absolute",
-                  right: 200,
+                  justifyContent: "center", // Centralized text and icon
                 }}
               >
                 Simpan
